lsm303agr split main
